I dont like it!! What f**k have I done?!

+++ srcs/raycast/sprites_manage.c
@@ -6,7 +6,7 @@
 /*   By: ssacrist <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/12/26 22:50:27 by ssacrist          #+#    #+#             */
-/*   Updated: 2021/01/17 21:37:53 by ssacrist         ###   ########.fr       */
+/*   Updated: 2021/01/18 00:03:18 by ssacrist         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,8 +23,8 @@ static void		allocate_sprite(t_cub3d *a)
 static t_sprite	calc_distance_nd_stature(t_cub3d *a, t_sprite sprite)
 {
 	sprite.distance = hypot(
-				((int)a->rayc.xray + 0.5) - a->rayc.xplyr,
-				((int)a->rayc.yray + 0.5) - a->rayc.yplyr)
+				a->rayc.xray + 0.5 - a->rayc.xplyr,
+				a->rayc.yray + 0.5 - a->rayc.yplyr)
 				* cos(a->rayc.anglray - a->rayc.dirplyr);
 	sprite.stature = a->fconf.xrendersize / sprite.distance;
 	sprite.init = round(a->fconf.yrendersize / 2.0 - sprite.stature / 2);
